+++ src/main/java/vboled/netcracker/musicstreamer/service/UserServiceImpl.java
@@ -52,12 +52,12 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public User read(int id) throws NoSuchElementException {
+    public User read(Long id) throws NoSuchElementException {
         return userRepository.findById(id).get();
     }
 
     @Override
-    public boolean update(User user, int id) {
+    public boolean update(User user, Long id) {
         if (userRepository.existsById(id)) {
             checkEmail(user.getEmail());
             checkPhone(user.getPhoneNumber());
@@ -71,7 +71,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public boolean delete(int id) {
+    public boolean delete(Long id) {
         if (userRepository.existsById(id)) {
             userRepository.deleteById(id);
             return true;
@@ -80,7 +80,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public boolean updateEmail(String email, int id) {
+    public boolean updateEmail(String email, Long id) {
         if (userRepository.existsById(id)) {
             User user = userRepository.getById(id);
             user.setEmail(email);
@@ -91,7 +91,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public boolean updatePhone(String phone, int id) {
+    public boolean updatePhone(String phone, Long id) {
         if (userRepository.existsById(id)) {
             User user = userRepository.getById(id);
             user.setPhoneNumber(phone);
@@ -102,7 +102,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public boolean updatePassword(String password, int id) {
+    public boolean updatePassword(String password, Long id) {
         if (userRepository.existsById(id)) {
             User user = userRepository.getById(id);
             user.setPassword(password);
@@ -169,13 +169,13 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public User userNotFullUpdate(User update, int id) {
+    public User userNotFullUpdate(User update, Long id) {
         User updated = updateCommonFields(update, id);
         userRepository.save(updated);
         return updated;
     }
 
-    private User updateCommonFields(User update, int id) throws UsernameNotFoundException, IllegalArgumentException {
+    private User updateCommonFields(User update, Long id) throws UsernameNotFoundException, IllegalArgumentException {
         if (!userRepository.existsById(id))
             throw new UsernameNotFoundException("No user with such id");
         User userToUpdate = userRepository.getById(id);
